% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMzPeak-functions.R,
%   R/MsBackendMzPeak.R
\name{MsBackendMzPeak}
\alias{MsBackendMzPeak}
\alias{MsBackendMzPeak-class}
\alias{show,MsBackendMzPeak-method}
\alias{[,MsBackendMzPeak-method}
\title{\code{Spectra} MS backend to read data from files in mzPeak format}
\usage{
MsBackendMzPeak()

\S4method{show}{MsBackendMzPeak}(object)

\S4method{[}{MsBackendMzPeak}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{object}{An \code{MsBackendMzPeak} object.}

\item{x}{AN \code{MsBackendMzPeak} object.}

\item{i}{For \code{[}: \code{integer} or \code{logical} defining to which spectra the data
should be subset.}

\item{j}{For \code{[}: ignored.}

\item{...}{additional arguments.}

\item{drop}{For \code{[}: ignored.}
}
\description{
The \code{MsBackendMzPeak} is an implementation of the \code{\link[Spectra:MsBackend]{Spectra::MsBackend()}}
class for \code{\link[Spectra:Spectra]{Spectra::Spectra()}} objects adding support for MS data import
from the \href{https://github.com/HUPO-PSI/mzPeak}{HUPO-PSI mzPeak} file format.
\code{MsBackendMzPeak} is an \emph{on-disk} implementation keeping only minimal
information in memory while loading any data (peaks data or spectra metadata)
from the original mzPeak files on-the-fly.
}
\details{
The \code{MsBackendMzPeak} backend uses functionality from the
\href{https://github.com/HUPO-PSI/mzpeak/R/mzpeak}{\emph{mzpeak}} R package for
accessing data in mzPeak files.

The \code{MsBackendMzPeak} extends the \code{\link[Spectra:MsBackendCached]{Spectra::MsBackendCached()}}\verb{backend allowing thus to intermediately store spectrum metadata (*spectra variables*) in a}data.frame` within the object (i.e., \strong{in memory}).
}
\section{Creation of backend objects}{


New \code{MsBackendMzPeak} backend objects can be created with the
\code{MsBackendMzPeak()} function.
}

\section{Subset, merge and filter data}{

\itemize{
\item \code{[}: subset a \code{MsBackendMzPeak} to the provided spectra (parameter \code{i}).
Internally, this will subset and re-order the cached data as well as the
indices and mzPeak files. The data in the original mzPeak files is not
affected.
}
}

\section{Data access}{


Data access functionality.
}

\author{
Johannes Rainer, Gabriele Tomè
}
